fix: Refactor number handlers to fix precision and overflow issues
- Split `ft_handle_int` logic into `ft_print_int` helper to separate calculation from printing.
- Split `ft_handle_nbrbase` logic into `ft_print_hex` helper to handle precision correctly.
- Switch to `long long` casting in `ft_handle_int` to prevent overflow with `INT_MIN`.
- Fix flag priority: explicitly disable zero padding when precision is present.
- Correct printing order: ensure sign and hex prefix (`0x`) appear before zero padding.
- Implement missing precision handling for `%u`, `%x`, and `%X`.
- Handle edge case where value is `0` with precision `.0` (print nothing).

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: vbleskin <vbleskin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/15 13:54:19 by vbleskin          #+#    #+#             */
-/*   Updated: 2025/12/07 18:34:35 by vbleskin         ###   ########.fr       */
+/*   Updated: 2025/12/07 20:36:21 by vbleskin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,7 +43,7 @@ t_struct	*ft_init_list(void);
 void		ft_reset_list(t_struct *list);
 int			ft_strlen(const char *str);
 const char	*ft_strchr(const char *str, char c);
-int			ft_nbrlen(int nb);
+int			ft_nbrlen(long long nb);
 int			ft_nbaselen(unsigned long long n, int base_len);
 int			ft_isdigit(int c);
 int			ft_isupper(const char *str);
